+++ core/src/main/scala/pureconfig/package.scala
@@ -7,9 +7,10 @@
 import java.io.{ OutputStream, PrintStream }
 import java.nio.file.{ Files, Path }
 
-import com.typesafe.config.{ ConfigFactory, Config => TypesafeConfig }
+import com.typesafe.config.{ Config => TypesafeConfig }
 import pureconfig.error.{ ConfigReaderException, ConfigReaderFailures, ConfigValueLocation }
 import pureconfig.ConfigConvert.improveFailures
+import pureconfig.backend.ConfigFactoryWrapper.{ invalidateCaches, load, loadFile, parseFile }
 
 import scala.reflect.ClassTag
 
@@ -23,8 +24,8 @@ package object pureconfig {
    *         isn't possible
    */
   def loadConfig[Config](implicit conv: ConfigConvert[Config]): Either[ConfigReaderFailures, Config] = {
-    ConfigFactory.invalidateCaches()
-    loadConfig[Config](ConfigFactory.load())(conv)
+    invalidateCaches()
+    load().right.flatMap(loadConfig[Config](_)(conv))
   }
 
   /**
@@ -36,8 +37,8 @@ package object pureconfig {
    *         isn't possible
    */
   def loadConfig[Config](namespace: String)(implicit conv: ConfigConvert[Config]): Either[ConfigReaderFailures, Config] = {
-    ConfigFactory.invalidateCaches()
-    improveFailures[Config](loadConfig[Config](ConfigFactory.load().getConfig(namespace))(conv), namespace, None)
+    invalidateCaches()
+    load().right.flatMap(rawConfig => improveFailures[Config](loadConfig[Config](rawConfig.getConfig(namespace))(conv), namespace, None))
   }
 
   /**
@@ -49,8 +50,8 @@ package object pureconfig {
    *         isn't possible
    */
   def loadConfig[Config](path: Path)(implicit conv: ConfigConvert[Config]): Either[ConfigReaderFailures, Config] = {
-    ConfigFactory.invalidateCaches()
-    loadConfig[Config](ConfigFactory.load(ConfigFactory.parseFile(path.toFile)))(conv)
+    invalidateCaches()
+    loadFile(path).right.flatMap(loadConfig[Config](_)(conv))
   }
 
   /**
@@ -63,9 +64,8 @@ package object pureconfig {
    *         isn't possible
    */
   def loadConfig[Config](path: Path, namespace: String)(implicit conv: ConfigConvert[Config]): Either[ConfigReaderFailures, Config] = {
-    ConfigFactory.invalidateCaches()
-    improveFailures[Config](
-      loadConfig[Config](ConfigFactory.load(ConfigFactory.parseFile(path.toFile)).getConfig(namespace))(conv), namespace, None)
+    invalidateCaches()
+    loadFile(path).right.flatMap(rawConfig => improveFailures[Config](loadConfig[Config](rawConfig.getConfig(namespace))(conv), namespace, None))
   }
 
   /** Load a configuration of type `Config` from the given `Config` */
@@ -87,8 +87,8 @@ package object pureconfig {
    *         isn't possible
    */
   def loadConfigWithFallback[Config](conf: TypesafeConfig)(implicit conv: ConfigConvert[Config]): Either[ConfigReaderFailures, Config] = {
-    ConfigFactory.invalidateCaches()
-    loadConfig[Config](conf.withFallback(ConfigFactory.load()))
+    invalidateCaches()
+    load().right.flatMap(rawConfig => loadConfig[Config](conf.withFallback(rawConfig)))
   }
 
   /**
@@ -101,8 +101,8 @@ package object pureconfig {
    *         isn't possible
    */
   def loadConfigWithFallback[Config](conf: TypesafeConfig, namespace: String)(implicit conv: ConfigConvert[Config]): Either[ConfigReaderFailures, Config] = {
-    ConfigFactory.invalidateCaches()
-    loadConfig[Config](conf.withFallback(ConfigFactory.load()), namespace)
+    invalidateCaches()
+    load().right.flatMap(rawConfig => loadConfig[Config](conf.withFallback(rawConfig), namespace))
   }
 
   private def getResultOrThrow[Config](failuresOrResult: Either[ConfigReaderFailures, Config])(implicit ct: ClassTag[Config]): Config = {
@@ -130,9 +130,11 @@ package object pureconfig {
    */
   @throws[ConfigReaderException[_]]
   def loadConfigOrThrow[Config](namespace: String)(implicit conv: ConfigConvert[Config], ct: ClassTag[Config]): Config = {
-    ConfigFactory.invalidateCaches()
-    val config = ConfigFactory.load().getConfig(namespace)
-    getResultOrThrow[Config](improveFailures[Config](loadConfig[Config](config)(conv), namespace, ConfigValueLocation(config.root())))
+    invalidateCaches()
+    val rawConfig = getResultOrThrow(load()).getConfig(namespace)
+    getResultOrThrow[Config](
+      improveFailures[Config](
+        loadConfig[Config](rawConfig)(conv), namespace, ConfigValueLocation(rawConfig.root())))
   }
 
   /**
@@ -143,9 +145,9 @@ package object pureconfig {
    */
   @throws[ConfigReaderException[_]]
   def loadConfigOrThrow[Config](path: Path)(implicit conv: ConfigConvert[Config], ct: ClassTag[Config]): Config = {
-    ConfigFactory.invalidateCaches()
-    val config = ConfigFactory.load(ConfigFactory.parseFile(path.toFile))
-    getResultOrThrow[Config](loadConfig[Config](config)(conv))
+    invalidateCaches()
+    val rawConfig = getResultOrThrow(loadFile(path))
+    getResultOrThrow[Config](loadConfig[Config](rawConfig)(conv))
   }
 
   /**
@@ -157,9 +159,11 @@ package object pureconfig {
    */
   @throws[ConfigReaderException[_]]
   def loadConfigOrThrow[Config](path: Path, namespace: String)(implicit conv: ConfigConvert[Config], ct: ClassTag[Config]): Config = {
-    ConfigFactory.invalidateCaches()
-    val config = ConfigFactory.load(ConfigFactory.parseFile(path.toFile)).getConfig(namespace)
-    getResultOrThrow[Config](improveFailures[Config](loadConfig[Config](config)(conv), namespace, ConfigValueLocation(config.root())))
+    invalidateCaches()
+    val rawConfig = getResultOrThrow(loadFile(path)).getConfig(namespace)
+    getResultOrThrow[Config](
+      improveFailures[Config](
+        loadConfig[Config](rawConfig)(conv), namespace, ConfigValueLocation(rawConfig.root())))
   }
 
   /**
@@ -194,8 +198,9 @@ package object pureconfig {
    */
   @throws[ConfigReaderException[_]]
   def loadConfigWithFallbackOrThrow[Config](conf: TypesafeConfig)(implicit conv: ConfigConvert[Config], ct: ClassTag[Config]): Config = {
-    ConfigFactory.invalidateCaches()
-    getResultOrThrow[Config](loadConfig[Config](conf.withFallback(ConfigFactory.load())))
+    invalidateCaches()
+    val rawConfig = getResultOrThrow(load())
+    getResultOrThrow[Config](loadConfig[Config](conf.withFallback(rawConfig)))
   }
 
   /**
@@ -207,8 +212,9 @@ package object pureconfig {
    */
   @throws[ConfigReaderException[_]]
   def loadConfigWithFallbackOrThrow[Config](conf: TypesafeConfig, namespace: String)(implicit conv: ConfigConvert[Config], ct: ClassTag[Config]): Config = {
-    ConfigFactory.invalidateCaches()
-    getResultOrThrow[Config](loadConfig[Config](conf.withFallback(ConfigFactory.load()), namespace))
+    invalidateCaches()
+    val rawConfig = getResultOrThrow(load())
+    getResultOrThrow[Config](loadConfig[Config](conf.withFallback(rawConfig), namespace))
   }
 
   /**
@@ -252,15 +258,19 @@ package object pureconfig {
    *
    * @param files Files ordered in decreasing priority containing part or all of a `Config`. Must not be empty.
    */
-  def loadConfigFromFiles[Config: ConfigConvert](files: Traversable[java.io.File]): Either[ConfigReaderFailures, Config] = {
+  def loadConfigFromFiles[Config: ConfigConvert](files: Traversable[Path]): Either[ConfigReaderFailures, Config] = {
     if (files.isEmpty) {
       ConfigConvert.failWithThrowable[Config](new IllegalArgumentException("The config files to load must not be empty."))(None)
     } else {
-      val resolvedTypesafeConfig = files
-        .map(ConfigFactory.parseFile)
-        .reduce(_.withFallback(_))
-        .resolve
-      loadConfig[Config](resolvedTypesafeConfig)
+      files
+        .map(parseFile)
+        .foldLeft[Either[ConfigReaderFailures, Seq[TypesafeConfig]]](Right(Seq())) {
+          case (c1, c2) =>
+            ConfigConvert.combineResults(c1, c2)(_ :+ _)
+        }
+        .right
+        .map(_.reduce(_.withFallback(_)).resolve)
+        .flatMap(loadConfig[Config])
     }
   }
 
@@ -269,8 +279,7 @@ package object pureconfig {
    * @return the configuration
    */
   @throws[ConfigReaderException[_]]
-  def loadConfigFromFilesOrThrow[Config: ConfigConvert](files: Traversable[java.io.File])(implicit ct: ClassTag[Config]): Config = {
+  def loadConfigFromFilesOrThrow[Config: ConfigConvert](files: Traversable[Path])(implicit ct: ClassTag[Config]): Config = {
     getResultOrThrow[Config](loadConfigFromFiles[Config](files))
   }
-
 }